requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
pytest==8.4.0
pytest-cov==6.0.0
selenium==4.33.0